ENG-739 Move color selector from Canvas tab to General tab
